[ADD] setting module
