+++ js/settings/localVariable.mjs
@@ -1,0 +1,7 @@
+const copySetting = localStorage.getItem("copy");
+const sizeSetting = localStorage.getItem("emojiSize");
+const recentSetting = localStorage.getItem("recentNum");
+const windowSetting = localStorage.getItem("windowSetting");
+const windowState = localStorage.getItem("windowState");
+
+export { copySetting, sizeSetting, recentSetting, windowSetting, windowState };
